% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vapply_1l}
\alias{vapply_1l}
\alias{vapply_1i}
\alias{vapply_1c}
\title{Safe Apply Wrappers}
\usage{
vapply_1l(x, fun, ...)

vapply_1i(x, fun, ...)

vapply_1c(x, fun, ...)
}
\arguments{
\item{x}{Passed as is to argument \code{X} of \code{\link[base:lapply]{*apply()}}.}

\item{fun}{Passed as is to argument \code{FUN} of \code{\link[base:lapply]{*apply()}}.}

\item{...}{Further arguments passed as is to \code{\link[base:lapply]{*apply()}}.
It later passes them to \code{fun}.}
}
\value{
\code{\link[=vapply_1l]{vapply_1l()}},
\code{\link[=vapply_1l]{vapply_1l()}}, and
\code{\link[=vapply_1c]{vapply_1c()}} respectively return a logical, an integer, and a character
vector having the same length as \code{x}. Names are always discarded.
}
\description{
These functions wrap a function of the \code{\link[base:lapply]{*apply()}}
family and enforce various values for convenience and additional
safety.
}
\seealso{
Other utility functions: 
\code{\link{split_ul}()},
\code{\link{stops}()},
\code{\link{strip_empty_strings}()},
\code{\link{to_string}()},
\code{\link{utc}()}
}
\concept{utility functions}
\keyword{internal}
