% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-block.R
\name{block}
\alias{block}
\alias{Block}
\alias{is_block}
\alias{format.Block}
\alias{print.Block}
\title{Translations Blocks}
\usage{
block(
  hash = "",
  text = "",
  text_key = "",
  locations = list(),
  translations = character()
)

is_block(x)

\method{format}{Block}(x, ...)

\method{print}{Block}(x, ...)
}
\arguments{
\item{hash}{A non-empty and non-\link[base:NA]{NA} character string. A
reproducible hash generated from \code{text} and used as a unique identifier
for the underlying \code{\link{Block}} object. It is typically generated by
a known \link[=get_hash_algorithms]{hashing algorithm}.}

\item{text}{A non-empty and non-\link[base:NA]{NA} character string. The source
text to be translated.}

\item{text_key}{A non-empty and non-\link[base:NA]{NA} character string. The
(default) language key of \code{text}. See \link{Translations Source Files} for
more information.}

\item{locations}{A list of \code{\link{Location}} objects referencing
\code{text} in the project. Since \code{text} can be reused as is multiple times,
it may have multiple locations.}

\item{translations}{A non-empty named character vector of non-\link[base:NA]{NA}
values. Values are actual translations of \code{text}. Their names correspond
to the underlying language keys. See \link{Translations Source Files} for more
information.}
}
\value{
\code{\link[=block]{block()}} returns a named list of class \code{\link{Block}} containing the
following elements:

\describe{
\item{\code{hash}}{See argument \code{hash}.}
\item{\code{text}}{See argument \code{text}.}
\item{\code{text_key}}{See argument \code{text_key}.}
\item{\code{locations}}{See argument \code{locations}.}
\item{\code{translations}}{See argument \code{translations}.}
\item{\code{.texts}}{A named character vector constructed from \code{text},
\code{text_key}, and \code{translations}. It combines all three elements into
a single vector and is equivalent to \code{c(text_key = text, translations)}.}
}

\code{\link[=is_block]{is_block()}} returns a logical.

\code{\link[=format.Block]{format.Block()}} returns a character string.

\code{\link[=print.Block]{print.Block()}} returns argument \code{x} invisibly.
}
\description{
An internal class that encapsulates translations of the same source text.
}
\note{
\code{\link[=block]{block()}} does \strong{not} generate hashes. This must be done \emph{before} creating
instances of the class. This is because \code{\link{Block}} is just a semantic
container.
}
\keyword{internal}
