% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-token.R
\name{token}
\alias{token}
\alias{tsf_block_line_token}
\title{Create Tokens}
\usage{
token(type = NULL, value = NULL, ..., .super = NULL)

tsf_block_line_token <- function(
  type = c(
    "NULL",
    "TITLE_HASH",
    "TITLE_KEY_SRC",
    "TITLE_KEY_TXT",
    "LOC_SRC_PATH",
    "LOC_SRC_RNG",
    "TXT"),
  value   = "",
  subtype = NULL
)
}
\arguments{
\item{type}{Any \R object identifying what type of data \code{value} holds.}

\item{value}{Any \R object. \code{\link[=tsf_block_line_token]{tsf_block_line_token()}} expects a character
string representing a source line extracted from a
\link[=Translations Source Files]{Translations Source File}.}

\item{...}{Further arguments. This is useful when extending class
\code{\link{Token}}.}

\item{.super}{An optional character vector of super-class names. This
is useful when extending class \code{\link{Token}}.}

\item{subtype}{An optional character string. If left as \code{NULL}, it is
automatically set equal to \code{type} for convenience.}
}
\value{
\code{\link[=token]{token()}} returns a named list of class \code{Token} containing the values of
arguments \code{type} and \code{value}. It will further contain fields that were
passed to \code{...}.

\code{\link[=tsf_block_line_token]{tsf_block_line_token()}} returns a named list of class \code{BlockLineToken}
containing the values of argument \code{type}, \code{value}, and \code{subtype}.
}
\description{
Helper classes to represent smallest units of meaningful data. A
\code{\link{Token}} object has a given \code{type} and an optional \code{subtype} that
may be used to parse the underlying \code{value}.
}
\seealso{
\code{\link[=tokenize_tsf_block_v1]{tokenize_tsf_block_v1()}},
\code{\link[=tokenize_tsf_block_line_v1]{tokenize_tsf_block_line_v1()}}
}
\keyword{internal}
