% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-string.R
\name{TranslatableString}
\alias{TranslatableString}
\alias{as.list.TranslatableString}
\alias{as.character.TranslatableString}
\alias{as.TranslatableString}
\title{Translatable strings}
\usage{
\method{as.list}{TranslatableString}(x, ...)

\method{as.character}{TranslatableString}(x, ...)

as.TranslatableString(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{\link[=as.TranslatableString]{as.TranslatableString()}} is an S3 generic function returning a
\link{TranslatableString} object.
}
\description{
Mark strings as being translatable, sanitize, and hash them.
}
\seealso{
\code{\link[=translate]{translate()}},
\code{\link[=as.TranslatableString.TranslateCall]{as.TranslatableString.TranslateCall()}}
}
\author{
Jean-Mathieu Potvin (\href{mailto:jeanmathieupotvin@ununoctium.dev}{jeanmathieupotvin@ununoctium.dev})
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{string}}{a sanitized character of length 1.}

\item{\code{signature}}{a character of length 1. The
\href{https://en.wikipedia.org/wiki/SHA-2}{SHA-256}
hash of field \verb{$string}.}

\item{\code{tokens}}{values previously passed to method
\code{\link[=TranslatableString]{$new()}} and coerced to a character.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TranslatableString-new}{\code{TranslatableString$new()}}
\item \href{#method-TranslatableString-format}{\code{TranslatableString$format()}}
\item \href{#method-TranslatableString-print}{\code{TranslatableString$print()}}
\item \href{#method-TranslatableString-cat}{\code{TranslatableString$cat()}}
\item \href{#method-TranslatableString-asCharacter}{\code{TranslatableString$asCharacter()}}
\item \href{#method-TranslatableString-asList}{\code{TranslatableString$asList()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranslatableString-new"></a>}}
\if{latex}{\out{\hypertarget{method-TranslatableString-new}{}}}
\subsection{Method \code{new()}}{
Create a \link{TranslatableString} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranslatableString$new(..., concat = " ")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{values to be concatenated into a single character string.}

\item{\code{concat}}{a character string used to concatenate values passed
to \code{...}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \link[R6:R6Class]{R6} object of class \link{TranslatableString}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranslatableString-format"></a>}}
\if{latex}{\out{\hypertarget{method-TranslatableString-format}{}}}
\subsection{Method \code{format()}}{
Format this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranslatableString$format(signatureLength = 64L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signatureLength}}{an integer between \code{0} and \code{64}. Desired
length of \verb{$signature}.
\itemize{
\item Leave it at \code{64} to use its full value.
\item Pass \code{0} to omit it from \code{header}.
\item Pass any other valid value to use its \code{signatureLength} first
characters.
}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A named character vector of length 2 containing the
following elements:

\describe{
\item{\code{header}}{a character string containing the
\code{signatureLength} first characters of field \verb{$signature}.}
\item{\code{string}}{current value of field \verb{$string}.}
}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranslatableString-print"></a>}}
\if{latex}{\out{\hypertarget{method-TranslatableString-print}{}}}
\subsection{Method \code{print()}}{
Print this object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranslatableString$print(signatureLength = 64L)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{signatureLength}}{passed to method \code{\link[=TranslatableString]{$format()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
This object invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranslatableString-cat"></a>}}
\if{latex}{\out{\hypertarget{method-TranslatableString-cat}{}}}
\subsection{Method \code{cat()}}{
Print unquoted and formatted value of field \verb{$string}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranslatableString$cat()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Nothing is printed if \verb{$string} is empty.
}

\subsection{Returns}{
This object invisibly.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranslatableString-asCharacter"></a>}}
\if{latex}{\out{\hypertarget{method-TranslatableString-asCharacter}{}}}
\subsection{Method \code{asCharacter()}}{
Coerce this object to a character.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranslatableString$asCharacter()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A single character. Current value of field \verb{$string}. It has
an attribute named \code{signature} corresponding to field \verb{$signature}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TranslatableString-asList"></a>}}
\if{latex}{\out{\hypertarget{method-TranslatableString-asList}{}}}
\subsection{Method \code{asList()}}{
Coerce this object to a list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TranslatableString$asList()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A named list. Each element is a field of this object.
See section \strong{Active bindings} above.
}
}
}
