% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-source.R
\name{find_source_in_file}
\alias{find_source_in_file}
\alias{find_source_in_exprs}
\alias{parse_file}
\title{Placeholder}
\usage{
find_source_in_file(
  path = "",
  encoding = "UTF-8",
  strict = FALSE,
  hash_algorithm = get_hash_algorithms()
)

find_source_in_exprs(
  .tokens = list(),
  .path = "",
  .strict = FALSE,
  .hash_algorithm = get_hash_algorithms()
)

parse_file(path = "", encoding = "UTF-8")
}
\description{
Placeholder.
}
\keyword{internal}
