% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-translations.R
\name{find_translations}
\alias{find_translations}
\title{Placeholder}
\usage{
find_translations(
  path = getwd(),
  encoding = "UTF-8",
  hash_algorithm = get_hash_algorithms()
)
}
\description{
Placeholder for future function \code{\link[=find_translations]{find_translations()}}.
}
