% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_vector}
\alias{format_vector}
\title{Format Vectors}
\usage{
format_vector(
  x = vector(),
  .top_label = NULL,
  .indent = 2L,
  .show_nokey = TRUE,
  .level = 1L
)
}
\arguments{
\item{x}{A \link[=vector]{vector}. It cannot be empty.}

\item{.top_label}{A \code{NULL} or a non-empty and non-\link[base:NA]{NA} character
string used as some kind of descriptive message. Its exact meaning depends
on the underlying context. In recursive calls to \code{\link[=format_vector]{format_vector()}}, this
is set equal to the name of the current element of \code{x}. \strong{It is not
validated for efficiency.}}

\item{.indent}{A non-\link[base:NA]{NA} integer value. Number of space characters
to use by \code{.levels} for indentation purposes. \strong{It is not validated for
efficiency.}}

\item{.show_nokey}{A non-\link[base:NA]{NA} logical value. Should \code{NULL} and/or
empty names be replaced by \code{"<nokey>"}? \strong{It is not validated for
efficiency.}}

\item{.level}{A non-\link[base:NA]{NA} integer value. The current nesting level
in recursive calls to \code{\link[=format_vector]{format_vector()}}. \strong{It is not validated for
efficiency.}}
}
\value{
A character vector.
}
\description{
Format atomic \link[=vector]{vectors} and/or \emph{recursive} structures such as
\link[=list]{lists}.
}
\details{
The intent of \code{\link[=format_vector]{format_vector()}} is to format any \link[=vector]{vector} into a
set of name/value pairs (given as character strings), where indentation
provides information on embedded structures. The following template shows
how it formats \code{x}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{<.top_label>:
  <x1_name>: <x1_value>
  <x2_name>: <x2_value>
  <x3_name>:
    <x3_x1_name>: <x3_x1_value>
    ...
  <nokey>: <xi_value>
  <nokey>: <xj_value>
  <nokey>:
    <nokey>: <xk_value>
    ...
}\if{html}{\out{</div>}}

The \code{"<nokey>"} strings won't be printed unless \code{show_nokey} is \code{TRUE}. If
it is \code{FALSE}, no label is printed for missing keys (names that are either
\code{NULL} or empty).
}
\note{
\code{\link[=format_vector]{format_vector()}} is the \emph{workhorse} function of all S3 \code{\link[=format]{format()}} methods:
\code{\link[=format.Translator]{format.Translator()}}, \code{\link[=format.Block]{format.Block()}}, and \code{\link[=format.Location]{format.Location()}}.
}
\examples{
x <- list(
    FirstName = "John",
    LastName  = "Doe",
    Address   = list(
        StreetAddress = "123 Main Street",
        City          = "Montreal",
        Province      = "Quebec",
        PostalCode    = "H0H 0H0"),
    Notes = c(
        "Send mail to",
        "address above."))

cat(transltr:::format_vector(x, "<JohnDoe>"), sep = "\n")
cat(transltr:::format_vector(x, "<JohnDoe>", .show_nokey = FALSE), sep = "\n")

}
\seealso{
Other utility functions: 
\code{\link{split_ul}()},
\code{\link{stops}()},
\code{\link{utc}()},
\code{\link{vapply_1l}()}
}
\concept{utility functions}
\keyword{internal}
