% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from-tsf.R
\name{from_tsf}
\alias{from_tsf}
\alias{split_tsf}
\alias{from_src_header}
\alias{from_src_blocks}
\alias{from_src_header_v1}
\alias{from_src_blocks_v1}
\alias{tokenize_src_block_v1}
\alias{from_src_block_v1}
\alias{from_src_block_title_v1}
\alias{from_src_block_txt_v1}
\alias{from_src_block_loc_v1}
\alias{from_src_block_loc_path_v1}
\alias{from_src_block_loc_range_v1}
\alias{src_block_line_token}
\title{Placeholder}
\usage{
from_tsf(x = character())

split_tsf(x = character())

from_src_header(x = character())

from_src_blocks(blocks = list(), template_version = get_template_versions())

from_src_header_v1(
  template_version = 1L,
  generated_by     = get_generated_by(),
  generated_on     = get_generated_on(),
  hash_algorithm   = get_hash_algorithms(),
  hash_length      = 32L,
  language_keys    = list(en = "English"),
  ...
)

from_src_blocks_v1(blocks = list())

tokenize_src_block_v1(x = character())

from_src_block_v1(tokens = list())

from_src_block_title_v1(token = src_block_line_token("TITLE_HASH"))

from_src_block_txt_v1(tokens = list())

from_src_block_loc_v1(tokens = list())

from_src_block_loc_path_v1(token = src_block_token("LOC_SRC_PATH"))

from_src_block_loc_range_v1(token = src_block_token("LOC_SRC_RNG"))

src_block_line_token(
  type = c("NULL", "TITLE_HASH", "TITLE_KEY_SRC", "TITLE_KEY_TXT", "LOC_SRC_PATH",
    "LOC_SRC_RNG", "TXT"),
  value = "",
  subtype = NULL
)
}
\description{
Placeholder for future documentation.
}
\seealso{
\link{Translations Source Files},
\code{\link[=read_translations]{read_translations()}},
\code{\link[=write_translations]{write_translations()}},
\code{\link[=extract_src_header]{extract_src_header()}},
\code{\link[=from_src_header]{from_src_header()}}
}
\concept{translations source files mechanisms}
\keyword{internal}
