% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-translations-from-expression.R
\name{getTranslationsFromExpression}
\alias{getTranslationsFromExpression}
\alias{.getTranslationsFromExpression}
\title{Find translations in an expression}
\usage{
getTranslationsFromExpression(expr = expression())

.getTranslationsFromExpression(expr = expression(), env = new.env())
}
\arguments{
\item{expr}{an \link[base:expression]{expression}. It can contain
\link[base:name]{symbols},
\link[base:call]{calls},
further \link[base:expression]{expressions}, etc. See \code{\link[base:parse]{base::parse()}} for
information.}

\item{env}{an \link[base:environment]{environment}.}
}
\value{
A named list (possibly empty) containing \link{TranslatableString} objects.
Names correspond to their signatures.
}
\description{
Extract strings passed to \code{\link[=translate]{translate()}} in an \link[base:expression]{expression}
vector.
}
\details{
Strings extracted from the same \link[base:call]{call} (\code{expr}) are coerced
to a single \link{TranslatableString} object.
}
\note{
The author was initially inspired by the source code of \code{\link[tools:xgettext]{tools::xgettext()}}.
Many thanks to the original author and to \verb{R Core Team}.
}
\section{Implementation}{


\code{\link[=getTranslationsFromExpression]{getTranslationsFromExpression()}} is split into:
\enumerate{
\item an outer part, \code{\link[=getTranslationsFromExpression]{getTranslationsFromExpression()}}, and
\item an inner part, \code{\link[=.getTranslationsFromExpression]{.getTranslationsFromExpression()}}.
}

The inner part encapsulates the \emph{main} recursion mechanism that loops
through expressions, modifying \code{env} (in place) each time it encounters
a \link[base:call]{call} to \code{\link[=translate]{translate()}}. The outer part initialiazes this
mechanism and returns the final output when it is done.
}

\seealso{
\link{TranslatableString},
\code{\link[tools:xgettext]{tools::xgettext()}}
}
\author{
Jean-Mathieu Potvin (\href{mailto:jeanmathieupotvin@ununoctium.dev}{jeanmathieupotvin@ununoctium.dev})
}
\keyword{internal}
