% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash-algorithms.R
\name{hash_algorithms}
\alias{hash_algorithms}
\title{Hashing algorithms}
\usage{
hash_algorithms()
}
\value{
A character vector of length 2.
}
\description{
\code{\link{transltr}} uses hashing algorithms to create globally unique
and reproducible identifiers from source texts. Two methods are available.
\subsection{\code{SHA1}: Secure Hash Algorithm 1}{

Method \code{sha1} corresponds to SHA-1 (Secure Hash Algorithm version 1), a
cryptographic hashing function. While it is now superseded by more secure
variants (SHA-256, SHA-512, etc.), it is still useful for non-sensitive
purposes. It is fast, collision-resistant, and may handle very large inputs.
It emits strings of 40 hexadecimal characters.
}

\subsection{\code{UTF8}: Cumulative UTF-8 Sum}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{This method is experimental. Use with caution.}

Method \code{utf8} is a simple method derived from cumulative sums of UTF-8 code
points (converted to integers). It is slightly faster than method \code{sha1} for
small inputs, and emits shorter hashes of a variable number of digits. The
length is porportional to the underlying input's length.

Strictly speaking, it is not a hashing algorithm per se. Instead, it
should be viewed as an identification algorithm that is highly likely to
produce different values for different inputs.
}
}
\note{
Further methods such as \href{https://github.com/Cyan4973/xxHash}{\code{xxhash}} will
become available in the future.
}
\examples{
hash_algorithms()

}
\seealso{
\code{\link{Translator}},
\code{\link{Text}}
}
\keyword{internal}
