% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{.hook_load}
\alias{.hook_load}
\alias{.hook_unload}
\title{Package hooks}
\usage{
.hook_load()

.hook_unload()
}
\value{
A \code{NULL} invisibly. Hooks are used for their side-effect.
\itemize{
\item \code{\link[=.hook_load]{.hook_load()}} creates an empty environment named
\code{\link{.__STR_ATTACHED_DB}} and attaches it to the
\link[=search]{search path}.
\item \code{\link[=.hook_unload]{.hook_unload()}} detaches and removes it.
}
}
\description{
These hooks are called just after loading and unloading the namespace of
\code{\link{transltr}}. See \code{\link[=getHook]{getHook()}} for details.
}
\details{
\code{\link[=.hook_load]{.hook_load()}} and \code{\link[=.hook_unload]{.hook_unload()}} are internally wrapped by \R's standard
hook functions \code{\link[=.onLoad]{.onLoad()}} and \code{\link[=.onUnload]{.onUnload()}} (as usual). The former are kept
separate for documentation and testing purposes.
}
\note{
Calling \code{\link[=.hook_unload]{.hook_unload()}} manually will break core mechanisms of the package.
}
\keyword{internal}
