% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-text.R
\name{read_text}
\alias{read_text}
\alias{write_text}
\title{Low-level text input and output}
\usage{
read_text(path = "", encoding = "UTF-8")

write_text(x, path = "")
}
\arguments{
\item{path}{A character string. A path to a file to read text from
or write text to.}

\item{encoding}{A character string. The file's underlying encoding.
In almost all cases, this should be \code{UTF-8}. Text supplied in other
encodings is internally re-encoded to \code{UTF-8} for portability.}

\item{x}{A character vector. Lines of text to write.}
}
\value{
\code{\link[=read_text]{read_text()}} returns a character vector.

\code{\link[=write_text]{write_text()}} is not yet implemented and throws an error.
}
\description{
\code{\link[=read_text]{read_text()}} and \code{\link[=write_text]{write_text()}} respectively wrap \code{\link[base:readLines]{base::readLines()}} and
\code{\link[base:writeLines]{base::writeLines()}}. They further validate their arguments, normalize
file paths, and re-encode inputs to \code{UTF-8} before reading and writing.
}
\keyword{internal}
