% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-translations.R
\name{read_translations}
\alias{read_translations}
\alias{write_translations}
\title{Read and Write Translations Source Files}
\usage{
read_translations(path = "", encoding = "UTF-8")

write_translations(x, path = "")
}
\arguments{
\item{path}{A character string. A path to a file to read text from
or write text to.}

\item{encoding}{A character string. The file's underlying encoding.
In almost all cases, this should be \code{UTF-8}. Text supplied in other
encodings is internally re-encoded to \code{UTF-8} for portability.}

\item{x}{An object of class \code{\link{Translator}}.}
}
\value{
\code{\link[=read_translations]{read_translations()}} returns a \code{\link{Translator}} object if \code{path}
points to a valid translations source file.

\code{\link[=write_translations]{write_translations()}} is not yet implemented and throws an error.
}
\description{
Read \link{Translations Source Files} and parse them as \code{\link{Translator}}
objects, or convert \code{\link{Translator}} objects back to
\link{Translations Source Files}.
}
\seealso{
\link{Translations Source Files}
}
