% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/language.R
\name{language_set}
\alias{language_set}
\alias{language_get}
\title{Current Language}
\usage{
language_set(key = "en")

language_get()
}
\arguments{
\item{key}{A non-empty and non-\link[base:NA]{NA} character string. A
language key.}
}
\value{
\code{\link[=language_set]{language_set()}} returns \code{NULL} invisibly. It is used for its side-effect
of setting environment variable \code{TRANSLTR_LANGUAGE}.

\code{\link[=language_get]{language_get()}} returns a character string (possibly empty). It corresponds
to the value of environment variable \code{TRANSLTR_LANGUAGE}.
}
\description{
Get or set the current language \strong{key} setting of \code{\link{transltr}}.
It is registered as an environment variable named \code{TRANSLTR_LANGUAGE}.
}
\details{
\code{\link[=language_set]{language_set()}} leaves the underlying locale as is. To change an \R
session's locale, use \code{\link[=Sys.setlocale]{Sys.setlocale()}} instead. Users can safely use
both simultaneously if required. See below for more information.
}
\note{
An environment variable is used because it can be shared among different
processes. This matters when using parallel and/or concurrent \R sessions.
}
\section{Locales versus languages}{

A \href{https://en.wikipedia.org/wiki/Locale_(computer_software)}{locale} is a
set of multiple low-level settings that relate to the user's language and
region. The \emph{language} itself is just one parameter among many others.

Modifying a locale on-the-fly \emph{can} be considered risky in some situations.
It may not be the optimal solution for merely changing textual representations
of a program or an application at runtime, as it may introduce unintended
changes and induce subtle bugs that are harder to fix.

Moreover, it makes sense for some applications and/or programs such as
\href{https://shiny.posit.co/}{Shiny applications} to decouple the front-end's
current language (what \emph{users} see) from the back-end's locale (what
\emph{developers} see). A UI may be displayed in a certain language while keeping
logs and \R internal \link[=message]{messages}, \link[=warning]{warnings}, and
\link[=stop]{errors} as is (in the source language).

Consequently, the language setting of \code{\link{transltr}} is purposely
kept separate from the underlying locale and removes the complexity of
having to support many of them. You may still do so while using the package.
}

