% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalize Text}
\usage{
normalize(...)
}
\arguments{
\item{...}{Any number of character vectors.}
}
\value{
A character string, possibly empty.
}
\description{
\code{\link[=normalize]{normalize()}} constructs a string from single-line and multi-line strings
passed to \code{...} All elements are (implicitly) coerced to character values
in the process.
}
\section{Text Normalization}{

\code{\link{transltr}} always normalizes character vectors. Elements passed
to \code{...} are transformed and assembled into a single character string to
ensure their consistency. Both single-line and multi-line strings are
supported.

Character vectors go through these five steps.
\enumerate{
\item Whitespaces (tabs, newlines, and repeated spaces) characters are
replaced by a single space.
\item Leading and trailing empty strings are discarded.
\item Empty strings inserted within non-empty elements of \code{...} are
interpreted as paragraph separators. They are replaced by two
newline characters (\verb{\\n\\n}).
\item All resulting elements (including paragraph separators) are coerced
into a single character string.
\item Any remaining leading and trailing whitespaces are stripped from the
character string.
}
}

\keyword{internal}
