% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{\%??\%}
\alias{\%??\%}
\title{Nullish coalescing operator}
\usage{
lhs \%??\% rhs
}
\arguments{
\item{lhs}{Any \R object. Left-hand side operand.}

\item{rhs}{Any \R object. Right-hand side operand.}
}
\value{
Argument \code{rhs} if \code{lhs} is \code{NULL}, and \code{lhs} otherwise.
}
\description{
The nullish coalescing (\verb{\%??\%}) operator is a logical operator that
returns its right-hand side operand when its left-hand side operand
is null and otherwise returns its left-hand side operand.
}
\note{
The actual description of the operator was taken from
\href{https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators/Nullish_coalescing}{mdn web docs}.
}
\seealso{
Other utility functions: 
\code{\link{vapply_1l}()}
}
\concept{utility functions}
\keyword{internal}
