% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-language.R
\name{set_language}
\alias{set_language}
\alias{get_language}
\title{Placeholder}
\usage{
set_language()

get_language()
}
\description{
Placeholder for future functions \code{\link[=set_language]{set_language()}} and \code{\link[=get_language]{get_language()}}.
}
