% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set-translator.R
\name{set_translator}
\alias{set_translator}
\alias{get_translator}
\title{Placeholder}
\usage{
set_translator(x = translator())

get_translator()
}
\description{
Placeholder for future functions \code{\link[=set_translator]{set_translator()}} and \code{\link[=get_translator]{get_translator()}}.
}
