% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_ul}
\alias{split_ul}
\title{Divide Into Groups}
\usage{
split_ul(...)
}
\arguments{
\item{...}{Potential arguments passed to \code{\link[base:split]{base::split()}}.}
}
\value{
A list. See \code{\link[base:split]{base::split()}} for further information.
}
\description{
\code{\link[=split_ul]{split_ul()}} wraps \code{\link[base:split]{base::split()}} and returns an \strong{u}nnamed \strong{l}ist.
}
\seealso{
Other utility functions: 
\code{\link{format_vector}()},
\code{\link{stops}()},
\code{\link{utc}()},
\code{\link{vapply_1l}()}
}
\concept{utility functions}
\keyword{internal}
