% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-block.R
\name{extract_src_blocks}
\alias{extract_src_blocks}
\alias{from_src_blocks}
\alias{from_src_block}
\alias{from_src_block_hash}
\alias{extract_src_translations}
\alias{from_src_translation}
\alias{from_src_language_key}
\title{Placeholder}
\usage{
extract_src_blocks(x = character())

from_src_blocks(src_blocks = list())

from_src_block(x = character())

from_src_block_hash(x = character(1L))

extract_src_translations(x = character())

from_src_translation(x = character())

from_src_language_key(x = character(1L))
}
\description{
Placeholder for future documentation.
}
\seealso{
\link{Translations Source Files},
\code{\link[=read_translations]{read_translations()}},
\code{\link[=write_translations]{write_translations()}},
\code{\link[=extract_src_header]{extract_src_header()}},
\code{\link[=from_src_header]{from_src_header()}}
}
\concept{source blocks mechanisms}
\keyword{internal}
