% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_chars}
\alias{strip_chars}
\title{Strip specific characters}
\usage{
strip_chars(x = character(), chars = character())
}
\arguments{
\item{x}{A character vector. It can be empty.}

\item{chars}{A character vector. It can be empty. If not, it may only contain
single-character elements.}
}
\value{
A character vector having the same length as \code{x}.
}
\description{
Strip specific (individual) characters from a character vector.
}
\examples{
transltr:::strip_chars("abc") # "abc"
transltr:::strip_chars(c("{{a}}", "`b`", "##  c"), c("{", "}", "`", "#", " ")) # c("a", "b", "c")

}
\seealso{
Other utility functions: 
\code{\link{\%??\%}()},
\code{\link{stops}()},
\code{\link{strip_empty_strings}()},
\code{\link{to_string}()},
\code{\link{utc}()},
\code{\link{vapply_1l}()}
}
\concept{utility functions}
\keyword{internal}
