% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_empty_strings}
\alias{strip_empty_strings}
\title{Strip leading and trailing empty strings}
\usage{
strip_empty_strings(x = character(), which = c("both", "leading", "trailing"))
}
\arguments{
\item{x}{A character vector. It can be empty or contain only empty elements.}

\item{which}{A character string equal to \code{both}, \code{leading}, or \code{trailing}.
What to remove from \code{x}.}
}
\value{
A character vector.
}
\description{
Strip leading and/or trailing superfluous empty elements from a character
vector, keeping only non-empty elements and empty strings intertwined with
them.
}
\note{
To strip all empty strings, see \code{\link[base:nchar]{base::nzchar()}}.
}
\examples{
transltr:::strip_empty_strings("") # character(0)

x <- c("", "", "a", "b", "", "c", "")

transltr:::strip_empty_strings(x)             # c("a", "b", "", "c")
transltr:::strip_empty_strings(x, "leading")  # c("a", "b", "", "c", "")
transltr:::strip_empty_strings(x, "trailing") # c("", "", "a", "b", "", "c", "")

}
\seealso{
Other utility functions: 
\code{\link{split_ul}()},
\code{\link{stops}()},
\code{\link{to_string}()},
\code{\link{utc}()},
\code{\link{vapply_1l}()}
}
\concept{utility functions}
\keyword{internal}
