% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag.R
\name{tag}
\alias{tag}
\alias{untag}
\alias{isTagged}
\title{Tag chararacter strings}
\usage{
tag(strings = character(), normalize = TRUE)

untag(strings = character(), keepDoubleQuotes = FALSE)

isTagged(strings = character())
}
\arguments{
\item{strings}{Character vector of non-\link[=NA]{base::NA} values. Empty elements
are left as is and are not tagged.}

\item{normalize}{A logical flag. Should existing braces be replaced?
Preceding and trailing outer spaces are removed. Escaped double quotes
are also removed.}

\item{keepDoubleQuotes}{A logical flag. Keep existing outer double quotes?}
}
\value{
\code{\link[=tag]{tag()}} and \code{\link[=untag]{untag()}} return a character vector of a length equal to the
length of argument \code{strings}.

\code{\link[=isTagged]{isTagged()}} returns a logical vector of a length equal to the length of
argument \code{strings}. A \code{TRUE} implies that the underlying element is
enclosed by double braces. You may use \code{\link[=tag]{tag()}} to normalize them.
}
\description{
Wrap character strings with double braces (\code{{{ }}}) and/or remove them.
}
\examples{
# Strings can be wrapped.
identical(tag("test string."), "{{ test string. }}")  ## TRUE

# Strings can be unwrapped.
identical(tag("{{ test string. }}"), "test string.")  ## TRUE

# Tags can be normalized.
identical(tag("  {{   test string.   }}   "), "{{ test string. }}")  ## TRUE

}
\author{
Jean-Mathieu Potvin (\href{mailto:jm@potvin.xyz}{jm@potvin.xyz})
}
