% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{text_read}
\alias{text_read}
\alias{text_write}
\title{Low-level Text Input and Output}
\usage{
text_read(path = "", encoding = "UTF-8")

text_write(x, path = "")
}
\arguments{
\item{path}{A character string. A path to a file to read text from,
or write text to.}

\item{encoding}{A non-empty and non-\link[base:NA]{NA} character string. The
underlying source character encoding. In almost all cases, this should
be UTF-8. Other encodings are internally re-encoded to UTF-8 for
portability.}

\item{x}{A character vector. Lines of text to write.}
}
\value{
\code{\link[=text_read]{text_read()}} returns a character vector.

\code{\link[=text_write]{text_write()}} is not yet implemented and throws an error.
}
\description{
\code{\link[=text_read]{text_read()}} and \code{\link[=text_write]{text_write()}} respectively wrap \code{\link[base:readLines]{base::readLines()}} and
\code{\link[base:writeLines]{base::writeLines()}}. They further validate their arguments, normalize
file paths, and re-encode inputs to \code{UTF-8} before reading and writing.
}
\keyword{internal}
