% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to-string.R
\name{to_string}
\alias{to_string}
\alias{to_string.default}
\title{Convert an R object to a character string}
\usage{
to_string(x, ...)

\method{to_string}{default}(x, quote_values = FALSE, last_sep = " or ", ...)
}
\arguments{
\item{x}{An object to be converted.}

\item{...}{Further arguments passed to or from other methods.}

\item{quote_values}{Should elements of \code{x} be quoted with single quotation
marks?}

\item{last_sep}{A character string that separates the last element from
the penultimate one.}
}
\value{
A character string constructed from \code{x}. Elements are separated
by \code{", "}, except for the last one. Argument \code{last_sep} is used for it.
}
\description{
This is an alternative to \code{\link[base:toString]{base::toString()}} that provides greater
flexiblity. It is used internally to construct more meaningful error
messages. It drops argument \code{width} because it is not useful in the
context of package \code{\link{transltr}}.
}
\seealso{
Other utility functions: 
\code{\link{\%??\%}()},
\code{\link{stops}()},
\code{\link{utc}()},
\code{\link{vapply_1l}()}
}
\concept{utility functions}
\keyword{internal}
