% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{isCallToTranslate}
\alias{isCallToTranslate}
\alias{newTranslateCall}
\alias{TranslateCall}
\alias{newTranslateCall.call}
\alias{newTranslateCall.default}
\alias{as.TranslatableString.TranslateCall}
\alias{doEvaluateTranslateDefaultArgs}
\title{Manipulate calls to translate()}
\usage{
isCallToTranslate(x)

newTranslateCall(x, ...)

\method{newTranslateCall}{call}(x, ...)

\method{newTranslateCall}{default}(x, ...)

\method{as.TranslatableString}{TranslateCall}(x, ...)

## Fetch, evaluate, and cache default arguments of translate()
TRANSLATE_DEFAULT_ARGS <- doEvaluateTranslateDefaultArgs()
}
\arguments{
\item{x}{an \R object.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\itemize{
\item \code{\link[=isCallToTranslate]{isCallToTranslate()}} returns a logical value.
\item \code{\link[=newTranslateCall]{newTranslateCall()}} and related methods return a \link{TranslateCall} object.
\item \code{\link[=doEvaluateTranslateDefaultArgs]{doEvaluateTranslateDefaultArgs()}} returns a named list holding the
default arguments of \code{\link[=translate]{translate()}}. Dot dot dot is removed.
}
}
\description{
Create \link[base:call]{calls} to \code{\link[=translate]{translate()}}, identify when this function
is called, flag calls as such, and convert them into \link{TranslatableString}
objects.
}
\details{
\code{\link[=newTranslateCall]{newTranslateCall()}} creates unevaluated function \link[base:call]{calls} to
\code{\link[=translate]{translate()}}. The default method can be used to construct such calls from
arbitrary arguments as if they were passed to \code{\link[=translate]{translate()}}. It considers
\code{x} to be the first argument passed to \code{...}

\code{\link[=isCallToTranslate]{isCallToTranslate()}} will not check whether the current state of
the \link[base:search]{search path} yields the correct \code{\link[=translate]{translate()}}
function (when the namespace is omitted).

\code{\link[=doEvaluateTranslateDefaultArgs]{doEvaluateTranslateDefaultArgs()}} is called once whenever \pkg{transltr}
is loaded to fetch and evaluate default arguments of \code{\link[=translate]{translate()}}. They
are cached via constant \code{TRANSLATE_DEFAULT_ARGS}.
}
\author{
Jean-Mathieu Potvin (\href{mailto:jeanmathieupotvin@ununoctium.dev}{jeanmathieupotvin@ununoctium.dev})
}
\keyword{internal}
