% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate Text}
\usage{
translate(
  ...,
  key = language_get(),
  scope = NULL,
  source_key = "en",
  concat = " "
)
}
\arguments{
\item{...}{Any number of character vectors. The source text to translate.}

\item{key}{A non-empty and non-\link[base:NA]{NA} character string. A language
key, i.e. a \emph{code} for a language. For example, \code{"jp"} could be the key
of \code{"日本語"} (Japanese). While users retain full control over keys, it is
best to use language codes stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, and
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
To maximize portability and to ease shared usage of language keys,
\code{\link{transltr}} recommends using codes stemming from the former.}

\item{scope}{A \code{NULL}, or an optional non-empty and non-\link[base:NA]{NA}
character string. See \code{\link[=translator_set]{translator_set()}} for more information.}

\item{source_key}{A non-empty and non-\link[base:NA]{NA} character string. The
language key of the (untranslated) source text.}

\item{concat}{A non-empty and non-\link[base:NA]{NA} character string used to
concatenate values passed to \code{...}.}
}
\description{
Translate source text to a language given by \code{key}. This is an high-level
interface to \code{\link[=Translator]{Translator$translate()}}.
}
\note{
It is recommended to \strong{always} use \code{\link[=translate]{translate()}} instead of
\code{\link[=Translator]{Translator$translate()}}. The latter is not detected by
\code{\link[=find_translations]{find_translations()}}, while the former is.
}
\examples{
## Create a Translator. This would normally be done
## automatically by find_source() or read_translations().
my_translator <- translator(block("en", en = "Hello!", es = "¡Hola!"))

## Register it (implicitly, under 'global' scope).
translator_set(my_translator)

## Set current language.
language_set("es")

## Request translations.
translate("Hello!") ## Outputs "¡Hola!".
translate("Hello!", key = "en") ## Outputs "Hello!".

}
\seealso{
\code{\link{Translator}},
\code{\link[=translator_set]{translator_set()}},
\code{\link[=language_set]{language_set()}}
}
