% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transltr-params.R
\name{get_template_versions}
\alias{get_template_versions}
\alias{get_hash_algorithms}
\alias{get_generated_by}
\alias{get_generated_on}
\title{Parameters and Constants}
\usage{
get_template_versions()

get_hash_algorithms()

get_generated_by()

get_generated_on()
}
\value{
\code{\link[=get_template_versions]{get_template_versions()}} returns an integer vector.

\code{\link[=get_hash_algorithms]{get_hash_algorithms()}} returns a character vector.

\code{\link[=get_generated_by]{get_generated_by()}} returns a character string.

\code{\link[=get_generated_on]{get_generated_on()}} is just a semantic alias for \code{\link[=utc]{utc()}} and returns
whatever it returns.
}
\description{
Helper functions that return valid values for various parameters and
components of package \code{\link{transltr}}.
}
\section{Hashing algorithms}{

Package \code{\link{transltr}} offers two methods.
\itemize{
\item Method \code{sha1} corresponds to SHA-1 (Secure Hash Algorithm 1),
a cryptographic hashing function. While it is now superseded by more
secure variants (SHA-256, SHA-512, etc.), it is still useful for non-
sensitive purposes. It is fast, collision-resistant, and may handle
very large inputs. It outputs strings of 40 hexadecimal characters.
\item Method \code{utf8} is an \strong{experimental hashing method} derived from the
cumulative sum of code points. It is slightly faster that \code{sha1} for
smaller inputs, and outputs shorter integers (as strings) of variable
lengths porportional to the input's length. It should be reserved for
expert use only.
}
}

\seealso{
\code{\link[=utc]{utc()}}
}
\keyword{internal}
