% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utc}
\alias{utc}
\title{Get Current Date and Time}
\usage{
utc()
}
\value{
A character string.
}
\description{
Get the current Coordinated Universal Time (UTC).
}
\details{
The format is

\if{html}{\out{<div class="sourceCode r">}}\preformatted{"<Full Month> <Day>, <Year> @ <Hours>:<Minutes>:<Seconds> (UTC)".
}\if{html}{\out{</div>}}
}
\seealso{
Other utility functions: 
\code{\link{format_vector}()},
\code{\link{split_ul}()},
\code{\link{stops}()},
\code{\link{vapply_1l}()}
}
\concept{utility functions}
\keyword{internal}
