% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-create-parent-dir.R
\name{create_parent_dir}
\alias{create_parent_dir}
\title{Parent Directory}
\usage{
create_parent_dir(path = "")
}
\arguments{
\item{path}{A non-empty and non-\link[base:NA]{NA} character string. A path to
a file to read text from, or write text to.}
}
\value{
A character string. A \link[=normalizePath]{normalized} and absolute path to
the parent directory of \code{path}.
}
\description{
Check whether the parent directory of a normalized \code{path} exists, and if
it does not, attempt to create it.
}
\seealso{
Other utility functions: 
\code{\link{format_vector}()},
\code{\link{split_ul}()},
\code{\link{stops}()},
\code{\link{to_string}()},
\code{\link{utc}()},
\code{\link{vapply_1l}()}
}
\concept{utility functions}
\keyword{internal}
