% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{utils-is}
\alias{utils-is}
\title{Test for specific shapes and values}
\arguments{
\item{x}{any \R object to be tested.}

\item{min}{a single numeric value. Lower inclusive boundary for \code{x}. This
value is an integer for \code{\link[=isSingleIntInRange]{isSingleIntInRange()}}.}

\item{max}{a single numeric value. Upper inclusive boundary for \code{x}. This
value is an integer for \code{\link[=isSingleIntInRange]{isSingleIntInRange()}}.}
}
\value{
A single logical. \code{TRUE} is returned unless \code{x} fails the test.
}
\description{
Check if a value has a specific shape (\link[base:typeof]{type},
\link[=length]{length}) and/or value (\link[base:NA]{NAs}, boundaries, etc.).
}
\details{
\code{\link[=isSingleChar]{isSingleChar()}} returns \code{TRUE} for any non-NA character string (empty
or non-empty).

\code{\link[=isSingleIntInRange]{isSingleIntInRange()}} returns \code{TRUE} for any single, non-NA integer
value falling in the range \verb{[min, max]}.
}
\author{
Jean-Mathieu Potvin (\href{mailto:jeanmathieupotvin@ununoctium.dev}{jeanmathieupotvin@ununoctium.dev})
}
