% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert}
\alias{assert}
\alias{assertString}
\alias{assertExistingFile}
\title{Assert that a value is valid}
\usage{
assertString(x, name = deparse(substitute(x)))

assertExistingFile(x, name = deparse(substitute(x)))
}
\arguments{
\item{x}{An \R object to be tested.}

\item{name}{A character string. A value representing \code{x} that is used to
construct meaningful error messages. By default, this corresponds to the
underlying \link[base:name]{symbol} passed to \code{x}.}
}
\value{
Argument \code{x} is returned invisibly if valid.
}
\description{
Validate a value and throw an error if it is not.
}
\examples{
myValue <- 1

\dontrun{
assertString(myValue)
# Error: TypeError: `myValue` must be a single non-NA character value.

assertString(myValue, "otherName")
# Error: TypeError: `otherName` must be a single non-NA character value.
}

}
\seealso{
Other assertions: 
\code{\link{assert-backend}}
}
\author{
Jean-Mathieu Potvin (\href{mailto:jm@potvin.xyz}{jm@potvin.xyz})
}
\concept{assertions}
\keyword{internal}
