% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-block.R
\name{block}
\alias{block}
\alias{.block}
\alias{is_block}
\alias{format.Block}
\alias{print.Block}
\alias{c.Block}
\alias{merge_blocks}
\alias{as_block}
\alias{as_block.call}
\alias{as_block.character}
\alias{Block}
\title{Translations Blocks}
\usage{
block(source_lang = "", ..., hash_algorithm = get_hash_algorithms())

.block(
  source_lang = "",
  source_text = "",
  hash_algorithm = get_hash_algorithms(),
  langs = character(),
  texts = character(),
  locations = list()
)

is_block(x)

\method{format}{Block}(x, ...)

\method{print}{Block}(x, ...)

\method{c}{Block}(...)

merge_blocks(..., hash_algorithm = get_hash_algorithms())

as_block(x, ...)

\method{as_block}{call}(x, locations = list(), hash_algorithm = get_hash_algorithms(), ...)

\method{as_block}{character}(
  x,
  source_lang = "",
  locations = list(),
  hash_algorithm = get_hash_algorithms(),
  ...
)
}
\arguments{
\item{source_lang}{A non-empty and non-\link[base:NA]{NA} character string. The
language of the (untranslated) source text.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}

\item{...}{Usage depends on the underlying function.
\itemize{
\item Any number of \code{\link{Location}} objects and/or named character
strings for \code{\link[=block]{block()}} (in no preferred order).
\item Any number of \code{\link{Block}} objects for \code{\link[=merge_blocks]{merge_blocks()}} and S3
method \code{\link[=c]{c()}}.
\item Further arguments passed to or from other methods for \code{\link[=format]{format()}},
\code{\link[=print]{print()}}, and \code{\link[=as_block]{as_block()}}.
}}

\item{hash_algorithm}{A character string. The algorithm to use when hashing
source information for identification purposes. It must be a value returned
by \code{\link[=get_hash_algorithms]{get_hash_algorithms()}}.}

\item{source_text}{A non-empty and non-\link[base:NA]{NA} character string. The
(untranslated) source text.}

\item{langs}{A character vector of non-empty and non-\link[base:NA]{NA}
values. The languages of \code{texts}. See argument \code{source_lang} for
more information.}

\item{texts}{A character vector of of non-empty and non-\link[base:NA]{NA}
values. The translations of \code{source_text}.}

\item{locations}{A list of \code{\link{Location}} objects or a single
\code{\link{Location}} object.}

\item{x}{Any \R object. A \code{\link{Block}} object for \code{\link[=format]{format()}} and
\code{\link[=print]{print()}}.}
}
\value{
\code{\link[=block]{block()}} and \code{\link[=.block]{.block()}} return an \code{\link[R6:R6Class]{R6}} object of class
\code{\link{Block}}.

\code{\link[=is_block]{is_block()}} returns a logical.

\code{\link[=format]{format()}} returns a character.

\code{\link[=print]{print()}} returns argument \code{x} invisibly.

\code{\link[=c]{c()}} returns a \code{\link{Block}} object. It can only combine objects
having the exact same \code{hash}, which is equivalent to having the same
\code{hash_algorithm}, \code{source_lang}, and \code{source_text}. In that case, the
underlying translations and locations are combined into coherent sets.

\code{\link[=merge_blocks]{merge_blocks()}} returns a list of \code{\link{Block}} objects. It is
a generalized version of \code{\link[=c]{c()}} that handles \code{\link{Block}} objects
having different hash(es).

\code{\link[=as_block]{as_block()}} returns a \code{\link{Block}} object.
}
\description{
Structure and manipulate source text and its translations.
}
\details{
A \code{\link{Block}} is first and foremost a collection of translations of
a given source text (i.e. text extracted from \R source scripts). Each
translation is uniquely identified by a language.

It has two further components.
\enumerate{
\item A repducible hash derived from an hashing algorithm that can be used
to identify the source text and language.
\item One or more source locations. See class \code{\link{Location}} for more
information.
}

Users should never attempt to manipulate the information contained in a
\code{\link{Block}}. Consider using exported features instead.

\code{\link{Block}} objects may be created on-the-fly for testing purposes
with \code{\link[=block]{block()}}.
}
\examples{
## Create a Block object.
transltr:::block("en",
    transltr:::location("a", 1L, 2L, 3L, 4L),
    transltr:::location("a", 1L, 2L, 3L, 4L),
    transltr:::location("b", 5L, 6L, 7L, 8L),
    transltr:::location("c", c(9L, 10L), c(11L, 12L), c(13L, 14L), c(15L, 16L)),
    en = "Hello, world!",
    fr = "Bonjour, monde!",
    es = "¡Hola Mundo!",
    ja = "こんにちは世界！")

## Combine Blocks objects.
b1 <- transltr:::block("en",
    transltr:::location("a", 1L, 2L, 3L, 4L),
    en = "Hello, world!",
    fr = "Bonjour, monde!",
    es = "¡Hola Mundo!",
    ja = "こんにちは世界！")

b2 <- transltr:::block("en",
    transltr:::location("a", 5L, 6L, 7L, 8L),
    en     = "Hello, world!",
    fr     = "Bonjour, monde!",
    es     = "¡Hola Mundo!",
    `ja-2` = "こんにちは世界！")

b3 <- transltr:::block("fr",
    transltr:::location("c", 1L, 2L, 3L, 4L),
    en     = "Hello, world!",
    fr     = "Bonjour, monde!",
    `es-2` = "¡Hola Mundo!",
    `ja-2` = "こんにちは世界！")

c(b1, b2)
transltr:::merge_blocks(b1, b2, b3)


## ------------------------------------------------
## Method `Block$set_translation`
## ------------------------------------------------

## Registering source_lang and source_text.
blk <- transltr:::Block$new()
blk$set_translation("en", "Hello, world!")
blk$source_lang <- "en"

## ------------------------------------------------
## Method `Block$set_translations`
## ------------------------------------------------

blk <- transltr:::Block$new()
blk$set_translations(en = "Hello, world!", fr = "Bonjour, monde!")

## ------------------------------------------------
## Method `Block$rm_translation`
## ------------------------------------------------

## Removing source_lang and source_text.
blk <- transltr:::Block$new()
blk$set_translations(en = "Hello, world!", fr = "Bonjour, monde!")
blk$source_lang <- "en"

blk$source_lang <- "fr"
blk$rm_translation("en")
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{hash}}{A non-empty and non-\link[base:NA]{NA} character string. A
reproducible hash generated from \code{source_lang} and \code{source_text}
using the algorithm specified by \code{hash_algorithm}. It is used as
a unique identifier for the \code{\link{Block}} object.

This is a \strong{read-only} field. It is automatically updated
whenever fields \code{source_lang} and/or \code{hash_algorithm} are updated.}

\item{\code{hash_algorithm}}{A character string. The algorithm to use when hashing
source information for identification purposes. It must be a value returned
by \code{\link[=get_hash_algorithms]{get_hash_algorithms()}}.}

\item{\code{source_lang}}{A non-empty and non-\link[base:NA]{NA} character string. The
language of the (untranslated) source text.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}

\item{\code{source_text}}{A non-empty and non-\link[base:NA]{NA} character string. The
(untranslated) source text. This is a \strong{read-only} field.}

\item{\code{languages}}{A character vector. Registered language codes.
This is a \strong{read-only} field.}

\item{\code{translations}}{A named character vector. Registered
translations of \code{source_text}. Names correspond to \code{languages}.
This is a \strong{read-only} field.}

\item{\code{locations}}{A list of \code{\link{Location}} objects giving
the location(s) of \code{source_text} in the underlying project. This
is a \strong{read-only} field.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Block-new}{\code{Block$new()}}
\item \href{#method-Block-get_translation}{\code{Block$get_translation()}}
\item \href{#method-Block-set_translation}{\code{Block$set_translation()}}
\item \href{#method-Block-set_translations}{\code{Block$set_translations()}}
\item \href{#method-Block-set_locations}{\code{Block$set_locations()}}
\item \href{#method-Block-rm_translation}{\code{Block$rm_translation()}}
\item \href{#method-Block-rm_location}{\code{Block$rm_location()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-new"></a>}}
\if{latex}{\out{\hypertarget{method-Block-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{\link{Block}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$new(hash_algorithm = get_hash_algorithms())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash_algorithm}}{A character string. The algorithm to use when hashing
source information for identification purposes. It must be a value returned
by \code{\link[=get_hash_algorithms]{get_hash_algorithms()}}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{\link[R6:R6Class]{R6}} object of class \code{\link{Block}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-get_translation"></a>}}
\if{latex}{\out{\hypertarget{method-Block-get_translation}{}}}
\subsection{Method \code{get_translation()}}{
Extract a translation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$get_translation(lang = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{A non-empty and non-\link[base:NA]{NA} character string. The
requested language.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method can also be used to extract \code{source_text}.
}

\subsection{Returns}{
A character string. \code{NULL} is returned if \code{lang} is not
registered.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-set_translation"></a>}}
\if{latex}{\out{\hypertarget{method-Block-set_translation}{}}}
\subsection{Method \code{set_translation()}}{
Register a translation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$set_translation(lang = "", text = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{A non-empty and non-\link[base:NA]{NA} character string. The
requested language.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}

\item{\code{text}}{A non-empty and non-\link[base:NA]{NA} character string. A
translation, or a source text.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method is also used to register \code{source_lang} and
\code{source_text} \strong{before} setting them as such. See Examples below.
}

\subsection{Returns}{
A \code{TRUE} (invisibly).
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{## Registering source_lang and source_text.
blk <- transltr:::Block$new()
blk$set_translation("en", "Hello, world!")
blk$source_lang <- "en"
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-set_translations"></a>}}
\if{latex}{\out{\hypertarget{method-Block-set_translations}{}}}
\subsection{Method \code{set_translations()}}{
Register one or more translations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$set_translations(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Any number of named, non-empty, and non-\link[base:NA]{NA}
character strings.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method can be viewed as a vectorized version of
method \code{set_translation()}.
}

\subsection{Returns}{
A \code{TRUE} (invisibly).
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{blk <- transltr:::Block$new()
blk$set_translations(en = "Hello, world!", fr = "Bonjour, monde!")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-set_locations"></a>}}
\if{latex}{\out{\hypertarget{method-Block-set_locations}{}}}
\subsection{Method \code{set_locations()}}{
Register one or more locations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$set_locations(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Any number of \code{\link{Location}} objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method calls \code{\link[=merge_locations]{merge_locations()}} to merge all
values passed to \code{...} together with previously registered
\code{\link{Location}} objects. The underlying registered
paths and/or ranges won't be duplicated.
}

\subsection{Returns}{
A \code{TRUE} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-rm_translation"></a>}}
\if{latex}{\out{\hypertarget{method-Block-rm_translation}{}}}
\subsection{Method \code{rm_translation()}}{
Remove a registered translation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$rm_translation(lang = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{A non-empty and non-\link[base:NA]{NA} character string. The
requested language.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
You cannot remove a \code{lang} registered as \code{source_lang}.
You must assign a new value as \code{source_lang} before doing so.
}

\subsection{Returns}{
A logical (invisibly) indicating whether the operation
succeeded or not.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{## Removing source_lang and source_text.
blk <- transltr:::Block$new()
blk$set_translations(en = "Hello, world!", fr = "Bonjour, monde!")
blk$source_lang <- "en"

blk$source_lang <- "fr"
blk$rm_translation("en")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-rm_location"></a>}}
\if{latex}{\out{\hypertarget{method-Block-rm_location}{}}}
\subsection{Method \code{rm_location()}}{
Remove a registered location.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$rm_location(path = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{A non-empty and non-\link[base:NA]{NA} character string
identifying a \code{\link{Location}} object to be removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{TRUE} (invisibly).
}
}
}
