% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-block.R
\name{block}
\alias{block}
\alias{is_block}
\alias{format.Block}
\alias{print.Block}
\alias{c.Block}
\alias{merge_blocks}
\alias{as_block}
\alias{as_block.call}
\alias{Block}
\title{Source Text and Translations}
\usage{
block(source_lang = "", ..., hash_algorithm = get_hash_algorithms())

is_block(x)

\method{format}{Block}(x, ...)

\method{print}{Block}(x, ...)

\method{c}{Block}(...)

merge_blocks(..., hash_algorithm = get_hash_algorithms())

as_block(x, ...)

\method{as_block}{call}(x, location = list(), hash_algorithm = get_hash_algorithms(), ...)
}
\arguments{
\item{source_lang}{A non-empty and non-\link[base:NA]{NA} character string. The
language of the (untranslated) source text.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}

\item{...}{Usage depends on the underlying function.
\itemize{
\item Any number of \code{\link{Location}} objects, and/or named character
strings for \code{\link[=block]{block()}} (in no preferred order).
\item Any number of \code{\link{Block}} objects for \code{\link[=merge_blocks]{merge_blocks()}} and S3
method \code{\link[=c]{c()}}.
\item Further arguments passed to or from other methods for \code{\link[=format]{format()}},
\code{\link[=print]{print()}}, and \code{\link[=as_block]{as_block()}}.
}}

\item{hash_algorithm}{A character string equal to one of the values returned
by \code{\link[=get_hash_algorithms]{get_hash_algorithms()}}. The algorithm to use when hashing source
information for identification purposes.}

\item{x}{Any \R object.}

\item{location}{A \code{\link{Location}} object.}
}
\value{
\code{\link[=block]{block()}}, \code{\link[=c]{c()}}, and \code{\link[=as_block]{as_block()}} return an \code{\link[R6:R6Class]{R6}} object of
class \code{\link{Block}}.

\code{\link[=is_block]{is_block()}} returns a logical value.

\code{\link[=format]{format()}} returns a character vector.

\code{\link[=print]{print()}} returns argument \code{x} invisibly.

\code{\link[=merge_blocks]{merge_blocks()}} returns a list of (combined) \code{\link{Block}} objects.
}
\description{
Store, structure, and manipulate \strong{a single} source text and its
translations.
}
\details{
A \code{\link{Block}} object is a collection of translations of a \emph{source text}
(i.e. text extracted from \R source scripts). It exposes a set of methods
that can be used to safely manipulate the information it contains, but it
is unlikely to be useful in typical circumstances. See class
\code{\link{Translator}} instead.
\subsection{Combining Block Objects}{

\code{\link[=c]{c()}} can only combine \code{\link{Block}} objects having the same \code{hash}.
This is equivalent to having the same \code{hash_algorithm}, \code{source_lang},
and \code{source_text}. In that case, the underlying translations and
\code{\link{Location}} objects are combined, and a new object is returned.

\code{\link[=merge_blocks]{merge_blocks()}} is a generalized version of \code{\link[=c]{c()}} that handles any number
of \code{\link{Block}} objects having possibly different hashes. It can be
viewed as a vectorized version of \code{\link[=c]{c()}}.
}
}
\examples{
## Create a Block object.
block("en",
    location("a", 1L, 2L, 3L, 4L),
    location("a", 1L, 2L, 3L, 4L),
    location("b", 5L, 6L, 7L, 8L),
    location("c", c(9L, 10L), c(11L, 12L), c(13L, 14L), c(15L, 16L)),
    en = "Hello, world!",
    fr = "Bonjour, monde!",
    es = "¡Hola Mundo!",
    ja = "こんにちは世界！")

## Combine Blocks objects.
b1 <- block("en",
    location("a", 1L, 2L, 3L, 4L),
    en = "Hello, world!",
    fr = "Bonjour, monde!",
    es = "¡Hola Mundo!",
    ja = "こんにちは世界！")

b2 <- block("en",
    location("a", 5L, 6L, 7L, 8L),
    en     = "Hello, world!",
    fr     = "Bonjour, monde!",
    es     = "¡Hola Mundo!",
    `ja-2` = "こんにちは世界！")

b3 <- block("fr",
    location("c", 1L, 2L, 3L, 4L),
    en     = "Hello, world!",
    fr     = "Bonjour, monde!",
    `es-2` = "¡Hola Mundo!",
    `ja-2` = "こんにちは世界！")

c(b1, b2)
merge_blocks(b1, b2, b3)


## ------------------------------------------------
## Method `Block$set_translation`
## ------------------------------------------------

## Registering source_lang and source_text.
blk <- Block$new()
blk$set_translation("en", "Hello, world!")
blk$source_lang <- "en"

## ------------------------------------------------
## Method `Block$set_translations`
## ------------------------------------------------

blk <- Block$new()
blk$set_translations(en = "Hello, world!", fr = "Bonjour, monde!")

## ------------------------------------------------
## Method `Block$rm_translation`
## ------------------------------------------------

## Removing source_lang and source_text.
blk <- Block$new()
blk$set_translations(en = "Hello, world!", fr = "Bonjour, monde!")
blk$source_lang <- "en"

blk$source_lang <- "fr"
blk$rm_translation("en")
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{hash}}{A non-empty and non-\link[base:NA]{NA} character string. A
reproducible hash generated from \code{source_lang} and \code{source_text},
and by using the algorithm specified by \code{hash_algorithm}. It is
used as a unique identifier for the underlying \code{\link{Block}}
object.

This is a \strong{read-only} field. It is automatically updated
whenever fields \code{source_lang} and/or \code{hash_algorithm} are updated.}

\item{\code{hash_algorithm}}{A character string equal to one of the values returned
by \code{\link[=get_hash_algorithms]{get_hash_algorithms()}}. The algorithm to use when hashing source
information for identification purposes.}

\item{\code{source_lang}}{A non-empty and non-\link[base:NA]{NA} character string. The
language of the (untranslated) source text.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}

\item{\code{source_text}}{A non-empty and non-\link[base:NA]{NA} character string. The
(untranslated) source text. This is a \strong{read-only} field.}

\item{\code{languages}}{A character vector. Registered language
codes. This is a \strong{read-only} field. Use methods below
to update it.}

\item{\code{translations}}{A named character vector. Registered
translations of \code{source_text}, including the latter. Names
correspond to \code{languages}. This is a \strong{read-only} field.
Use methods below to update it.}

\item{\code{locations}}{A list of \code{\link{Location}} objects giving
the location(s) of \code{source_text} in the underlying project. It
can be empty. This is a \strong{read-only} field. Use methods below
to update it.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Block-new}{\code{Block$new()}}
\item \href{#method-Block-get_translation}{\code{Block$get_translation()}}
\item \href{#method-Block-set_translation}{\code{Block$set_translation()}}
\item \href{#method-Block-set_translations}{\code{Block$set_translations()}}
\item \href{#method-Block-set_locations}{\code{Block$set_locations()}}
\item \href{#method-Block-rm_translation}{\code{Block$rm_translation()}}
\item \href{#method-Block-rm_location}{\code{Block$rm_location()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-new"></a>}}
\if{latex}{\out{\hypertarget{method-Block-new}{}}}
\subsection{Method \code{new()}}{
Create a \code{\link{Block}} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$new(hash_algorithm = get_hash_algorithms())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{hash_algorithm}}{A character string equal to one of the values returned
by \code{\link[=get_hash_algorithms]{get_hash_algorithms()}}. The algorithm to use when hashing source
information for identification purposes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An \code{\link[R6:R6Class]{R6}} object of class \code{\link{Block}}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-get_translation"></a>}}
\if{latex}{\out{\hypertarget{method-Block-get_translation}{}}}
\subsection{Method \code{get_translation()}}{
Extract a translation, or the source text.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$get_translation(lang = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{A non-empty and non-\link[base:NA]{NA} character string. The
requested language.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A character string. \code{NULL} is returned if the requested
translation is not available.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-set_translation"></a>}}
\if{latex}{\out{\hypertarget{method-Block-set_translation}{}}}
\subsection{Method \code{set_translation()}}{
Register a translation, or the source text.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$set_translation(lang = "", text = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{A non-empty and non-\link[base:NA]{NA} character string. The
requested language.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}

\item{\code{text}}{A non-empty and non-\link[base:NA]{NA} character string. A
translation, or a source text.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method is also used to register \code{source_lang} and
\code{source_text} \strong{before} setting them as such. See Examples below.
}

\subsection{Returns}{
A \code{TRUE} (invisibly).
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{## Registering source_lang and source_text.
blk <- Block$new()
blk$set_translation("en", "Hello, world!")
blk$source_lang <- "en"
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-set_translations"></a>}}
\if{latex}{\out{\hypertarget{method-Block-set_translations}{}}}
\subsection{Method \code{set_translations()}}{
Register one or more translations, and/or the source
text.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$set_translations(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Any number of named, non-empty, and non-\link[base:NA]{NA}
character strings.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method can be viewed as a vectorized version of
method \code{set_translation()}.
}

\subsection{Returns}{
A \code{TRUE} (invisibly).
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{blk <- Block$new()
blk$set_translations(en = "Hello, world!", fr = "Bonjour, monde!")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-set_locations"></a>}}
\if{latex}{\out{\hypertarget{method-Block-set_locations}{}}}
\subsection{Method \code{set_locations()}}{
Register one or more locations.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$set_locations(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Any number of \code{\link{Location}} objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
This method calls \code{\link[=merge_locations]{merge_locations()}} to merge all
values passed to \code{...} together with previously registered
\code{\link{Location}} objects. The underlying registered
paths and/or ranges won't be duplicated.
}

\subsection{Returns}{
A \code{TRUE} (invisibly).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-rm_translation"></a>}}
\if{latex}{\out{\hypertarget{method-Block-rm_translation}{}}}
\subsection{Method \code{rm_translation()}}{
Remove a registered translation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$rm_translation(lang = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lang}}{A non-empty and non-\link[base:NA]{NA} character string
identifying a translation to be removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
You cannot remove \code{lang} when it is registered as the
current \code{source_lang}. You must update \code{source_lang} before
doing so.
}

\subsection{Returns}{
A logical (invisibly) indicating whether the operation
succeeded or not.
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{## Removing source_lang and source_text.
blk <- Block$new()
blk$set_translations(en = "Hello, world!", fr = "Bonjour, monde!")
blk$source_lang <- "en"

blk$source_lang <- "fr"
blk$rm_translation("en")
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Block-rm_location"></a>}}
\if{latex}{\out{\hypertarget{method-Block-rm_location}{}}}
\subsection{Method \code{rm_location()}}{
Remove a registered location.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Block$rm_location(path = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{A non-empty and non-\link[base:NA]{NA} character string
identifying a \code{\link{Location}} object to be removed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{TRUE} (invisibly).
}
}
}
