% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-translator.R
\name{translator}
\alias{translator}
\alias{is_translator}
\alias{Translator}
\title{Placeholder}
\usage{
translator()

is_translator()
}
\description{
Placeholder for future class \code{\link{Translator}}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Translator-clone}{\code{Translator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Translator-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Translator-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Translator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
