% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{.__LGL_DEBUG_FLAG}
\alias{.__STR_EMPTY_OBJ}
\alias{.__STR_UNDEFINED}
\alias{.__STR_NO_NAME}
\alias{.__STR_FORMAL_CONCAT_DEFAULT}
\alias{.__STR_FORMAL_SOURCE_LANG_DEFAULT}
\alias{.__CHR_EXCLUDED_NS}
\title{Internal Constants}
\format{
\itemize{
\item \code{.__LGL_DEBUG_FLAG} is a logical value always equal to \code{FALSE}.
It is only ever (temporarily) set equal to \code{TRUE} in unit tests via
\code{\link[testthat:local_mocked_bindings]{testthat::with_mocked_bindings()}}. It is used to force errors that
are hard to test with usual methods.
}

\itemize{
\item \code{.__STR_EMPTY_OBJ} is a character string equal to \code{"<none>"}. It
standardizes the format of empty objects.
}

\itemize{
\item \code{.__STR_UNDEFINED} is a character string equal to \code{"<unset>"}. It
signals that a field is not yet set, and may change how empty strings are
formatted.
}

\itemize{
\item \code{.__STR_NO_NAME} is a character string equal to \code{"<nokey>"}. It
is (sometimes) used when formatting empty names.
}

\itemize{
\item \code{.__STR_FORMAL_CONCAT_DEFAULT} is a character string equal to \code{" "}.
It matches the value of formal argument \code{concat} of \code{\link[=translate]{translate()}}.
}

\itemize{
\item \code{.__STR_FORMAL_SOURCE_LANG_DEFAULT} is a character string equal
to \code{"en"}. It matches the value of formal argument \code{source_key} of
\code{\link[=translate]{translate()}}.
}

\itemize{
\item \code{.__CHR_EXCLUDED_NS} is a character vector containing elements
\code{"base"}, and \code{"transltr"}. These are the names of excluded packages (or
\link[=loadNamespace]{namespaces}) when attempting to determine the scope of
a \code{\link{Translator}} object. See \code{\link[=translator_set]{translator_set()}} and
\code{\link[=translator_get]{translator_get()}} for more information.
}
}
\usage{
.__LGL_DEBUG_FLAG

.__STR_EMPTY_OBJ

.__STR_UNDEFINED

.__STR_NO_NAME

.__STR_FORMAL_CONCAT_DEFAULT

.__STR_FORMAL_SOURCE_LANG_DEFAULT

.__CHR_EXCLUDED_NS
}
\description{
These constants are used internally by \code{\link{transltr}}. They ensure
consistency across its features. Their prefix states their underlying type.

\strong{Users should never interact with them.}
}
\keyword{internal}
