% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{.__STR_EMPTY_OBJ}
\alias{.__STR_UNDEFINED}
\alias{.__STR_NO_NAME}
\alias{.__STR_FORMAL_CONCAT_DEFAULT}
\alias{.__STR_FORMAL_KEY_DEFAULT}
\alias{.__OBJ_TRANSLATOR}
\title{Internal Constants}
\format{
\itemize{
\item \code{.__STR_EMPTY_OBJ} is a character string equal to \code{"<none>"}.
}

\itemize{
\item \code{.__STR_UNDEFINED} is a character string equal to \code{"<unset>"}.
}

\itemize{
\item \code{.__STR_NO_NAME} is a character string equal to \code{"<nokey>"}.
}

\itemize{
\item \code{.__STR_FORMAL_CONCAT_DEFAULT} is a character string equal to \code{" "}.
}

\itemize{
\item \code{.__STR_FORMAL_KEY_DEFAULT} is a character string equal to \code{"en"}.
}

\itemize{
\item \code{.__OBJ_TRANSLATOR} is a character string equal to \code{".__translator__"}.
}
}
\usage{
.__STR_EMPTY_OBJ

.__STR_UNDEFINED

.__STR_NO_NAME

.__STR_FORMAL_CONCAT_DEFAULT

.__STR_FORMAL_KEY_DEFAULT

.__OBJ_TRANSLATOR
}
\description{
These constants are used internally by \code{\link{transltr}}. They ensure
consistency across its features. Their prefix states their underlying type.
}
\details{
\code{.__OBJ_TRANSLATOR} gives the name of the attribute that references the
latest registered \code{\link{Translator}} object. It is set by
\code{\link[=set_translator]{set_translator()}} and read by \code{\link[=get_translator]{get_translator()}}.
}
\note{
Constants \code{.__STR_FORMAL_CONCAT_DEFAULT} and \code{.__STR_FORMAL_KEY_DEFAULT} are
respectively equal to the default values of formal arguments \code{concat} and
\code{source_key} of \code{\link[=translate]{translate()}}.
}
\keyword{internal}
