% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constant}
\alias{constant}
\alias{.__LGL_DEBUG_FLAG}
\title{Internal Constants}
\format{
\code{.__LGL_DEBUG_FLAG} is an internal logical value always equal to \code{FALSE}
used to force errors that are otherwise hard, or (almost) impossible to
test otherwise. It is only ever (temporarily) set equal to \code{TRUE} via
\code{\link[testthat:local_mocked_bindings]{testthat::with_mocked_bindings()}} in \code{\link[testthat:test_that]{testthat::test_that()}} blocks. They
are additional unit tests in place to ensure it is always \code{FALSE} outside
of unit tests.
}
\usage{
constant(which = c("concat", "empty", "unset", "unknown", "untranslated"))

.__LGL_DEBUG_FLAG
}
\value{
A character string, or \code{NULL} if value passed to \code{which} is unavailable.
}
\description{
Fetch shared constants used by \code{\link{transltr}}. Constants ensure
consistency among all features of the package.
}
\examples{
transltr:::constant("concat")        ## Outputs " "
transltr:::constant("empty")         ## Outputs "<none>"
transltr:::constant("unset")         ## Outputs "<unset>"
transltr:::constant("unknown")       ## Outputs "<unknown>"
transltr:::constant("untranslated")  ## Outputs "# Erase this comment and provide a translation."

# NULL is returned if which has no corresponding entry.
transltr:::constant("__undefined__")

}
\keyword{internal}
