% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constant}
\alias{constant}
\alias{.__LGL_DEBUG_FLAG}
\title{Internal Constants}
\format{
\code{.__LGL_DEBUG_FLAG} is a logical value always equal to \code{FALSE}.
It is only ever (temporarily) set equal to \code{TRUE} in unit tests via
\code{\link[testthat:local_mocked_bindings]{testthat::with_mocked_bindings()}} for development, and testing purposes.
It is used to force errors that are otherwise hard, or impossible to test
via usual methods. They are unit tests in place that ensures official
releases are always shipped with this flag set to \code{FALSE}.
}
\usage{
constant(
  which = c(
    "generated-by",
    "concat",
    "empty",
    "placeholder",
    "unset",
    "unknown"
  )
)

.__LGL_DEBUG_FLAG
}
\value{
A character string, or \code{NULL} if value passed to \code{which} is unavailable.
}
\description{
Fetch shared constants used by \code{\link{transltr}}. Constants ensure
consistency among all features of the package.
}
\examples{
constant("generated-by")  ## Outputs "R package transltr 0.0.1"
constant("concat")        ## Outputs " "
constant("empty")         ## Outputs "<none>"
constant("placeholder")   ## Outputs "<none>"
constant("unset")         ## Outputs "<unset>"
constant("unknown")       ## Outputs "<unknown>"

# NULL is returned if which has no corresponding entry.
constant("undefined")

}
\seealso{
\code{\link[=hash_algorithms]{hash_algorithms()}}
}
\keyword{internal}
