% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{.__STR_EMPTY_OBJ}
\alias{.__STR_UNDEFINED}
\alias{.__STR_NO_NAME}
\alias{.__STR_FORMAL_CONCAT_DEFAULT}
\alias{.__STR_FORMAL_KEY_DEFAULT}
\alias{.__STR_ATTACHED_DB}
\title{Internal Constants}
\format{
\itemize{
\item \code{.__STR_EMPTY_OBJ} is a character string equal to \code{"<none>"}. It
standardizes the format of empty objects.
}

\itemize{
\item \code{.__STR_UNDEFINED} is a character string equal to \code{"<unset>"}. It
signals that a field is not yet set, and may change how empty strings are
formatted.
}

\itemize{
\item \code{.__STR_NO_NAME} is a character string equal to \code{"<nokey>"}. It
is (sometimes) used when formatting empty names.
}

\itemize{
\item \code{.__STR_FORMAL_CONCAT_DEFAULT} is a character string equal to \code{" "}.
It matches the value of formal argument \code{concat} of \code{\link[=translate]{translate()}}.
}

\itemize{
\item \code{.__STR_FORMAL_KEY_DEFAULT} is a character string equal to \code{"en"}.
It matches the value of formal argument \code{source_key} of \code{\link[=translate]{translate()}}.
}

\itemize{
\item \code{.__STR_ATTACHED_DB} is a character string equal to
\code{"transltr:translators"}. This is the name of the environment
\code{\link{transltr}} attaches to the \link[=search]{search path}
when it is (first) loaded. Registered \code{\link{Translator}}
objects are assigned there.
}
}
\usage{
.__STR_EMPTY_OBJ

.__STR_UNDEFINED

.__STR_NO_NAME

.__STR_FORMAL_CONCAT_DEFAULT

.__STR_FORMAL_KEY_DEFAULT

.__STR_ATTACHED_DB
}
\description{
These constants are used internally by \code{\link{transltr}}. They ensure
consistency across its features. Their prefix states their underlying type.

\strong{Users should never interact with these constants.}
}
\keyword{internal}
