% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash.R
\name{algorithms}
\alias{algorithms}
\title{Hashing Algorithms}
\usage{
algorithms()
}
\description{
These algorithms map a character string to another character string of
hexadecimal characters highly likely to be unique. The latter is used
to uniquely identify a source text (and the underlying source language).
}
\details{
\subsection{Secure Hash Algorithm 1}{

Method \code{sha1} corresponds to SHA-1 (Secure Hash Algorithm version 1), a
cryptographic hashing function. While it is now superseded by more secure
variants (SHA-256, SHA-512, etc.), it is still useful for non-sensitive
purposes. It is fast, collision-resistant, and may handle very large inputs.
It emits strings of 40 hexadecimal characters.
}

\subsection{Cumulative UTF-8 Sum}{

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

\strong{This method is experimental. Use with caution.}

Method \code{utf8} is a simple method derived from cumulative sums of UTF-8 code
points (converted to integers). It is slightly faster than method \code{sha1} for
small inputs and emits hashes with a width porportional to the underlying
input's length. It is used for testing purposes internally.
}
}
\keyword{internal}
