% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-translations.R
\name{find_translations}
\alias{find_translations}
\title{Placeholder}
\usage{
find_translations()
}
\description{
Placeholder for future function \code{\link[=find_translations]{find_translations()}}.
}
