% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find-translations.R
\name{find_translations}
\alias{find_translations}
\title{Placeholder}
\usage{
find_translations(
  path = getwd(),
  encoding = "UTF-8",
  hash_algorithm = get_hash_algorithms()
)
}
\arguments{
\item{path}{Placeholder.}

\item{encoding}{Placeholder.}

\item{hash_algorithm}{A character string equal to one of the values returned
by \code{\link[=get_hash_algorithms]{get_hash_algorithms()}}. The algorithm to use when hashing source
information for identification purposes.}
}
\description{
Placeholder.
}
