% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src-block.R
\name{extract_src_blocks}
\alias{extract_src_blocks}
\alias{tokenize_src_block}
\title{Placeholder}
\usage{
extract_src_blocks(x = character())

tokenize_src_block(x = character())
}
\description{
Placeholder for future documentation.
}
\seealso{
\link{Translations Source Files},
\code{\link[=read_translations]{read_translations()}},
\code{\link[=write_translations]{write_translations()}},
\code{\link[=extract_src_header]{extract_src_header()}},
\code{\link[=from_src_header]{from_src_header()}}
}
\concept{source blocks mechanisms}
\keyword{internal}
