% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stopf.R
\name{stopf}
\alias{stopf}
\title{Throw errors}
\usage{
stopf(type = c("TypeError"), fmt = character(1L), ...)
}
\arguments{
\item{type}{Character string stating the underlying error's type.}

\item{fmt, ...}{Passed to function \code{\link[base:sprintf]{sprintf()}}.}
}
\value{
Nothing. This function is used for its side-effect.
}
\description{
This function is \strong{experimental}.

\code{\link[=stopf]{stopf()}} combines \code{\link[base:sprintf]{sprintf()}} and
\code{\link[base:stop]{stop()}}. It further replaces usual \code{Error: ...}
prefix by an informative error \code{type} such as \code{TypeError: ...} .
}
\examples{
stopf("TypeError", "argument must contain exactly \%i values.", 7L)
# TypeError: argument must contain exactly 7 values.

}
\author{
Jean-Mathieu Potvin (\href{mailto:jm@potvin.xyz}{jm@potvin.xyz})
}
