% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Future work}
\usage{
translate(..., lang = "en", translator = NULL)
}
\arguments{
\item{...}{to be implemented.}

\item{lang}{to be implemented.}

\item{translator}{to be implemented.}
}
\description{
To be implemented.
}
\author{
Jean-Mathieu Potvin (\href{mailto:jeanmathieupotvin@ununoctium.dev}{jeanmathieupotvin@ununoctium.dev})
}
