% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate Text}
\usage{
translate(
  ...,
  lang = language_get(),
  scope = NULL,
  concat = constant("concat"),
  source_lang = language_source_get()
)
}
\arguments{
\item{...}{Any number of literal character vectors. The source text to
translate. Values can be empty and/or \link[base:NA]{NA}, but can lead to
unexpected results.}

\item{lang}{A non-empty and non-\link[base:NA]{NA} character string. The
underlying language.

A language is usually a code (of two or three letters) for a native
language name. While users retain full control over codes, it is best
to use language codes stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}

\item{scope}{A \code{NULL}, or an optional non-empty and non-\link[base:NA]{NA}
character string.}

\item{concat}{A non-empty and non-\link[base:NA]{NA} character string used to
concatenate values passed to \code{...}.}

\item{source_lang}{A non-empty and non-\link[base:NA]{NA} character string. The
language of the (untranslated) source text. See argument \code{lang} for more
information.}
}
\description{
Translate source text. This is an high-level interface to
\code{\link[=Translator]{Translator$translate()}}.
}
\details{
See \code{\link[=translator_set]{translator_set()}} for more information on scopes.

It is strongly recommended to always include the namespace when using
\code{\link[=translate]{translate()}}, i.e. \code{transltr::translate()}. Doing so ensures that there
will be no ambiguity at runtime.
}
\examples{
#' ## Set source language.
language_source_set("en")

## Create a Translator. This would normally be done
## automatically by find_source() or translator_read().
my_translator <- translator(text(en = "Hello!", es = "¡Hola!"))

## Register it (implicitly, under 'global' scope).
translator_set(my_translator)

## Set current language.
language_set("es")

## Request translations.
translate("Hello!") ## Outputs "¡Hola!".
translate("Hello!", lang = "en") ## Outputs "Hello!".

}
\seealso{
\code{\link{Translator}},
\code{\link[=translator_set]{translator_set()}},
\code{\link[=language_set]{language_set()}}
}
