% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Translate Text}
\usage{
translate(
  ...,
  lang = language_get(),
  scope = NULL,
  concat = " ",
  source_lang = "en"
)
}
\arguments{
\item{...}{Any number of literal character vectors. The source text to
translate. Values can be empty and/or \link[base:NA]{NA}, but can lead to
unexpected results.}

\item{lang}{A non-empty and non-\link[base:NA]{NA} character string. The
requested language.

A language is usually a code (of two or three letters) for a native
language name. For example, \code{ja} is a code for \code{日本語} (Japanese). While
users retain full control over codes, it is best to use language codes
stemming from well-known schemes such as
\href{https://en.wikipedia.org/wiki/IETF_language_tag}{IETF BCP 47}, or
\href{https://en.wikipedia.org/wiki/List_of_ISO_639_language_codes}{ISO 639-1}.
Doing so maximizes portability and cross-compatibility between packages.}

\item{scope}{A \code{NULL}, or an optional non-empty and non-\link[base:NA]{NA}
character string.}

\item{concat}{A non-empty and non-\link[base:NA]{NA} character string used to
concatenate values passed to \code{...}.}

\item{source_lang}{A non-empty and non-\link[base:NA]{NA} character string. The
language of the (untranslated) source text. See argument \code{lang} for more
information.}
}
\description{
Translate source text. This is an high-level interface to
\code{\link[=Translator]{Translator$translate()}}.
}
\details{
See \code{\link[=translator_set]{translator_set()}} for more information on scopes.
}
\examples{
## Create a Translator. This would normally be done
## automatically by find_source() or read_translations().
my_translator <- translator(block("en", en = "Hello!", es = "¡Hola!"))

## Register it (implicitly, under 'global' scope).
translator_set(my_translator)

## Set current language.
language_set("es")

## Request translations.
translate("Hello!") ## Outputs "¡Hola!".
translate("Hello!", lang = "en") ## Outputs "Hello!".

}
\seealso{
\code{\link{Translator}},
\code{\link[=translator_set]{translator_set()}},
\code{\link[=language_set]{language_set()}}
}
