% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate.R
\name{translate}
\alias{translate}
\title{Placeholder}
\usage{
translate(..., lang = get_language(), concat = " ", key = "en")
}
\description{
Placeholder for future function \code{\link[=translate]{translate()}}.
}
