% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translator.R
\name{translator_import}
\alias{translator_import}
\alias{translator_export}
\title{Read and Write Portable Translators}
\usage{
translator_import(path = "", encoding = "UTF-8")

translator_export(x, path = "")
}
\arguments{
\item{path}{A character string. A path to a file to read text from,
or write text to.}

\item{encoding}{A non-empty and non-\link[base:NA]{NA} character string. The
underlying source character encoding. In almost all cases, this should
be UTF-8. Other encodings are internally re-encoded to UTF-8 for
portability.}

\item{x}{An object of class \code{\link{Translator}}.}
}
\value{
\code{\link[=translator_import]{translator_import()}} returns a \code{\link{Translator}} object if \code{path}
points to a valid translations source file.

\code{\link[=translator_export]{translator_export()}} is not yet implemented and throws an error.
}
\description{
Read \link{Portable Translator Files} and parse them as \code{\link{Translator}}
objects, or convert \code{\link{Translator}} objects back to
\link{Portable Translators Files}.
}
\seealso{
\link{Portable Translators Files}
}
