% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transltr-package.R
\docType{package}
\name{transltr-package}
\alias{transltr}
\alias{transltr-package}
\title{transltr: A general purpose translator for R}
\description{
A collection of translation mechanisms. Detect messages to be be translated in an application, organize their corresponding translated messages, and seamlessly use them.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeanmathieupotvin/transltr}
  \item Report bugs at \url{https://github.com/jeanmathieupotvin/transltr/issues/new}
}

}
\author{
\strong{Maintainer}: Jean-Mathieu Potvin \email{jeanmathieupotvin@ununoctium.dev}

Other contributors:
\itemize{
  \item Jérôme Lavoué \email{jerome.lavoue@umontreal.ca} (\href{https://orcid.org/0000-0003-4950-5475}{ORCID}) [contributor, funder, reviewer]
}

}
\keyword{internal}
