% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transltr-package.R
\docType{package}
\name{transltr-package}
\alias{transltr}
\alias{transltr-package}
\title{transltr: A General Purpose Translator for R}
\description{
An internationalization framework that enables full support of many languages in any R application. Detect text to translate within a project. Import and manage translations with a user-friendly interface. Easily export source text that requires translation to a plain text format that fosters collaboration with other non-technical and external collaborators.
}
\section{Getting Started}{

Placeholder for future documentation.
}

\section{Bugs and Feedback}{

You may submit feedback, ideas, and bugs by creating an
\href{https://github.com/jeanmathieupotvin/transltr/issues/new}{issue on GitHub}.
Your help is appreciated.
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jeanmathieupotvin/transltr}
  \item Report bugs at \url{https://github.com/jeanmathieupotvin/transltr/issues/new}
}

}
\author{
\strong{Maintainer}: Jean-Mathieu Potvin \email{jeanmathieupotvin@ununoctium.dev}

Other contributors:
\itemize{
  \item Jérôme Lavoué \email{jerome.lavoue@umontreal.ca} (\href{https://orcid.org/0000-0003-4950-5475}{ORCID}) [contributor, funder, reviewer]
}

}
\keyword{internal}
