% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utc.R
\name{utc}
\alias{utc}
\title{Get Current Date and Time}
\usage{
utc()
}
\value{
A character string.
}
\description{
Get the current Coordinated Universal Time (UTC).
}
\details{
The format is

\if{html}{\out{<div class="sourceCode r">}}\preformatted{"<Full Month>, <Day of Month>, <Year> @ <Hours>:<Minutes>:<Seconds> (UTC)".
}\if{html}{\out{</div>}}
}
\note{
The \code{\link[=utc]{utc()}} function is defined in its own script because it has its
own test script. Keeping it separate from other utility functions is
better for readability.
}
\seealso{
Other utility functions: 
\code{\link{split_ul}()},
\code{\link{stops}()},
\code{\link{strip_empty_strings}()},
\code{\link{to_string}()},
\code{\link{vapply_1l}()}
}
\concept{utility functions}
\keyword{internal}
