% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-is.R
\name{isSingleChar}
\alias{isSingleChar}
\title{Internal tools: test for specific shapes and values}
\usage{
isSingleChar(x)
}
\arguments{
\item{x}{any \R object to be tested.}
}
\value{
A single logical. \code{TRUE} is returned unless \code{x} fails the underlying
test.
}
\description{
Check if a value has a specific shape (\link[base:typeof]{type},
\link[=length]{length}) and/or value (\link[base:NA]{NAs}, boundaries, etc.).
}
\details{
\code{\link[=isSingleChar]{isSingleChar()}} returns \code{TRUE} for any empty or non-empty character string.
}
\seealso{
Other internal tools: 
\code{\link{map}()},
\code{\link{stopf}()}
}
\author{
Jean-Mathieu Potvin (\href{mailto:jeanmathieupotvin@ununoctium.dev}{jeanmathieupotvin@ununoctium.dev})
}
\concept{internal tools}
\keyword{internal}
