% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trickRoxyygenTemporarily}
\alias{trickRoxyygenTemporarily}
\alias{\%??\%}
\title{Other miscellaneous tools}
\usage{
trickRoxyygenTemporarily()

lhs \%??\% rhs
}
\arguments{
\item{lhs, rhs}{any \R object.}
}
\value{
\verb{\%||\%} returns returns \code{rhs} whenever \code{lhs} evaluates to \code{NULL}.
}
\description{
Small miscelleneous functions.
}
\details{
\verb{\%||\%} is the usual null coalescing operator.
}
\author{
Jean-Mathieu Potvin (\href{mailto:jeanmathieupotvin@ununoctium.dev}{jeanmathieupotvin@ununoctium.dev})
}
\keyword{internal}
